Temporary workarounds a bug in callNextMethod()
The unique() S4 generic defined in BiocGenerics is defined with

  setGeneric("unique", signature="x")

so is different from the base::unique() implicit S4 generic defined by
packages like rJava or flexmix. Unfortunately, if rJava or flexmix is
loaded before BiocGenerics, then any call to callNextMethod() that happen
from within a method defined on the BiocGenerics::unique() generic will
look for the next method in the wrong generic, that is, it will look in
the base::unique() implicit generic instead of in the BiocGenerics::unique()
generic.
This breaks the "unique" method for List objects which contains a call to
callNextMethod(). For now the workaround is to replace this call with an
explicit call to the "unique" method for Vector objects.

The sort() S4 generic defined in BiocGenerics is defined with

  setGeneric("sort", signature="x")

so the "sort" method for List objects is also affected by this bug in
callNextMethod(). We use a similar workaround for this method too.