2 fixes and 1 enhancement to the "rbind" method for DataFrame objects
Most of the logic that used to be implemented in the "rbind" method for
DataFrame objects is now in a new "bindROWS" method for these objects.
The "rbind" method for DataFrame objects now is just a thin wrapper
around a call to bindROWS().

This new "bindROWS" method for DataFrame objects fixes the following long
standing bugs:
  1. rbind() now properly handles DataFrame objects with duplicated
     colnames. Note that the new behavior is consistent with
     base::rbind.data.frame().
  2. rbind() now properly handles DataFrame objects with columns that
     are 1D arrays.

It also provides the following enhancement:
  3. rbind() now supports DataFrame objects with the same column names
     but in different order, even when some of the column names are
     duplicated. How rbind() re-aligns the columns of the various objects
     to bind with those of the first object is consistent with what
     base:::rbind.data.frame() does.