drop as.env,DataTable
This method was carefully written to support both RangedData and
DataFrame. DataFrame is now supported by with,SimpleList. IRanges has
been updated to keep with,RangedData working. We need to think about the
role of DataTable in the framework. It is not clear whether other
derivatives would have worked with this method implementation.