coerce,ANY,DataFrame now **always** propagates the names
Like as.data.frame(), coerce,ANY,DataFrame was dropping the names of the
object to coerce in case there were duplicates. Now that DataFrame objects
support duplicate row names (since commit 6dd5e618), there is no more
reason to do that. So now the names are always propagated.
For example 'as(c(a=1, a=2), "DataFrame")' now propagates the names.