Fix as.data.frame() on nested DataFrames
Refactor the "as.data.frame" method for DataFrame objects. The new
implementation calls as.data.frame() on each column instead of
as.list(). This recursive approach allows the new implementation
to naturally support columns containing nested data-frame-like objects
and/or other complicated objects (as long as these complicated objects
in turn support as.data.frame()).