Simplify normalizeSingleBracketReplacementValue() interface
normalizeSingleBracketReplacementValue() is an internal generic function.
Its argument list so far was: value, x, i. However, last argument (i) has
no place in this list, and, as a matter of fact, none of the existing
normalizeSingleBracketReplacementValue() methods is using it. This change
gets rid of it so the new argument list is: value, x.