Add DFrame class
Just a direct DataFrame extension with no additional slot for now. Once all
serialized DataFrame instances are replaced with DFrame instances (which
will take several BioC release cycles) we'll be able to move the DataFrame
slots from the DataFrame class definition to the DFrame class definition.
The final goal is to have DataFrame become a virtual class with no slots
that only extends DataTable, and DFrame a concrete DataFrame and SimpleList
subclass that has the same slots as the current DataFrame class.