Override base::I() with a less broken I()
This is an ugly hack and hopefully it is temporary only! See
https://stat.ethz.ch/pipermail/r-devel/2020-October/080038.html for
the full story. Packages that use I() internally will automatically
pick up the fix if they already bulk import S4Vectors. Otherwise
they need to add:

    import(S4Vectors)

for bulk import (recommended), or:

    importFrom(S4Vectors,I)

for selective import, to their NAMESPACE file.