Drop classNameForDisplay() method for AsIs objects
Was added in IRanges in 2013 (commit 7b0ddea), then moved to S4Vectors
in 2014. In the meantime the DataFrame() constructor got modified to
drop the AsIs class from the supplied columns, so it's no longer helpful
or desirable to hide the fact that a column is an AsIs object. This means
that the classNameForDisplay() method for AsIs objects can go away.