Fix regression introduced at previous commit in internal helper bindROWS2()
This regression got introduced at commit ad90e78 (S4Vectors 0.32.3) and
was causing 'S4Vectors:::bindROWS2(x, list(y))' to not combine the levels
of factor-Rle objects 'x' and 'y' if one of the 2 objects has length 0.
For example, with:

  x <- Rle(factor(levels="a"))
  y <- Rle(factor("b"))

it was returning a factor-Rle with the same levels as 'y' and was ignoring
the levels of 'x'. With this fix we get:

  > S4Vectors:::bindROWS2(x, list(y))
  factor-Rle of length 1 with 1 run
    Lengths: 1
    Values : b
  Levels(2): a b

  > S4Vectors:::bindROWS2(y, list(x))
  factor-Rle of length 1 with 1 run
    Lengths: 1
    Values : b
  Levels(2): b a

As expected!

At a higher-level, this regression was breaking c() on GRanges objects
when one of the objects to combine has length 0 and a non-empty set of
seqlevels. See https://github.com/Bioconductor/GenomicRanges/issues/65
for the details.