Adapt test_Rle_runq_integer() to a change in R-devel's quantile()
The (undocumented) change in the behavior of quantile() has to do with
the class of the returned object when the input is an integer vector:

- In R < 4.3:
   > x <- sample(99, 20)
   > sapply(1:9, function(type) class(quantile(x, type=type)))
   [1] "integer" "numeric" "integer" "integer" "numeric" "numeric" "numeric"
   [8] "numeric" "numeric"

- In R >= 4.3:
   > sapply(1:9, function(type) class(quantile(x, type=type)))
   [1] "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric"
   [8] "numeric" "numeric"

So in R >= 4.3, the returned class is always numeric, whatever the input or
quantile algorithm specified via the 'type' argument.

This change broke the tests in test_Rle_runq_integer(). This commit adapts
the tests to make them work again with the new behavior of quantile().