Fix bug in internal helper bindROWS2()
This fixes a misbehavior in S4Vectors:::bindROWS2() that happens when the
first element to combine ('x') is an atomic vector and at least one of the
other objects to combine ('objects') is a List derivative. For example:

  x1 <- 1:3
  x2 <- IntegerList(41:45, 51:52)
  x3 <- 6:7
  S4Vectors:::bindROWS2(x1, list(x2, x3))

Before the fix, this was returning an ordinary list of length 6 where
the 4th list element is x2.

Now with the fix it's returning an IntegerList object of length 7, as
it should:

  S4Vectors:::bindROWS2(x1, list(x2, x3))
  # IntegerList of length 7
  # [[1]] 1
  # [[2]] 2
  # [[3]] 3
  # [[4]] 41 42 43 44 45
  # [[5]] 51 52
  # [[6]] 6
  # [[7]] 7